chore(ui): move HUDMenuButton to organisms
