+++ src/AGENTS_DEVELOPMENT_SUMMARY.md
@@ -37,7 +37,7 @@ The table below lists core modules grouped by network layer and their current st
 |Client|`client/ui/molecules/LevelGem.ts`|Usable|Shows current level|
 |Client|`client/ui/molecules/ProgressBar.ts`|Usable|Basic progress bar|
 |Client|`client/ui/molecules/BarMeter.ts`|Usable|Tokenized bar meter|
-|Client|`client/ui/molecules/HUDMenuButton.ts`|Usable|HUD toggle button|
+|Client|`client/ui/organisms/HUDMenuButton.ts`|Usable|HUD toggle button|
 |Client|`client/ui/molecules/AbilityInfoPanel.ts`|Usable|Ability description panel|
 |Client|`client/ui/molecules/UserMessage.ts`|Usable|Center-screen popup messages|
 |Client|`client/states/SettingsState.ts`|Usable|Reactive settings container|
