+++ src/client/ui/molecules/index.ts
@@ -15,7 +15,6 @@ export * from "./AbilityInfoPanel";
 export * from "./BarMeter";
 export * from "./CountdownTimer";
 export * from "./ExperienceBar";
-export * from "./HUDMenuButton";
 export * from "./LevelGem";
 export * from "./ProgressBar";
 export * from "./UserMessage";
